added review func

+++ course-service/src/main/java/com/sunny/microservices/course/entity/Review.java
@@ -17,6 +17,6 @@ public class Review extends BaseEntity{
     @Id
     String id;
     String userId;
-    Double rating;
+    Integer rating;
     String content;
 }
